pyproject.toml
[project]
name = "polymarket-trading-bot"
version = "0.1.0"
description = "Polymarket CLOB trading bot (arbitrage-first)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "bot146" }]
dependencies = [
  "requests>=2.31.0",
  "pydantic>=2.5.0",
  "python-dotenv>=1.0.0",
  "websocket-client>=1.7.0",
  "tenacity>=8.2.3",
  "rich>=13.7.0",
  "py-clob-client>=0.17.0"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.4",
  "pytest-asyncio>=0.23.3",
  "ruff>=0.1.14",
  "mypy>=1.8.0",
  "types-requests>=2.31.0.20240106",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
strict_optional = true
